added snapshots
