--- src/Controls/tests/TestCases.HostApp/FeatureMatrix/ProgressBar/ProgressBarOptionsPage.xaml
@@ -1,126 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:Maui.Controls.Sample"
-             x:Class="Maui.Controls.Sample.ProgressBarOptionsPage"
-             x:DataType="local:ProgressBarViewModel"
-             Title="ProgressBarFeature">
-       <ContentPage.ToolbarItems>
-              <ToolbarItem Text="Apply"
-                           Clicked="ApplyButton_Clicked"
-                           AutomationId="Apply"/>
-       </ContentPage.ToolbarItems>
-
-       <Grid HorizontalOptions="Center"
-             ColumnDefinitions="Auto, Auto"
-             RowDefinitions="Auto, Auto, Auto, Auto, Auto, Auto"
-             RowSpacing="10"
-             ColumnSpacing="15">
-
-              <!-- Progress -->
-              <Label Grid.Row="0"
-                     Grid.Column="0"
-                     Text="{Binding Progress, StringFormat='Progress: {0:F2}'}"
-                     VerticalOptions="Center"/>
-              <Entry Grid.Row="0"
-                     Grid.Column="1"
-                     x:Name="ProgressEntry"
-                     AutomationId="ProgressEntry"
-                     TextChanged="OnProgressChanged"
-                     Keyboard="Numeric"
-                     HorizontalOptions="Start"
-                     VerticalOptions="Center"
-                     WidthRequest="100"/>
-
-              <!-- BackgroundColor -->
-              <Label Grid.Row="1"
-                     Grid.Column="0"
-                     Text="Background Color:"
-                     VerticalOptions="Center"/>
-              <StackLayout Grid.Row="1"
-                           Grid.Column="1"
-                           Orientation="Horizontal"
-                           Spacing="5"
-                           VerticalOptions="Center">
-                     <Button Text="Orange"
-                             BackgroundColor="Orange"
-                             TextColor="White"
-                             Clicked="BackgroundColorButton_Clicked"
-                             AutomationId="BackgroundColorOrangeButton"/>
-                     <Button Text="Light Blue"
-                             BackgroundColor="LightBlue"
-                             TextColor="Black"
-                             Clicked="BackgroundColorButton_Clicked"
-                             AutomationId="BackgroundColorLightBlueButton"/>
-              </StackLayout>
-
-              <!-- IsVisible -->
-              <Label Grid.Row="2"
-                     Grid.Column="0"
-                     Text="Is Visible:"
-                     VerticalOptions="Center"/>
-              <RadioButton Grid.Row="2"
-                           Grid.Column="1"
-                           VerticalOptions="Center"
-                           AutomationId="IsVisibleFalseRadio"
-                           GroupName="IsVisible"
-                           Content="False"
-                           CheckedChanged="OnIsVisibleCheckedChanged"/>
-
-              <!-- Flow Direction -->
-              <Label Grid.Row="3"
-                     Grid.Column="0"
-                     Text="Flow Direction:"
-                     FontSize="15"
-                     VerticalOptions="Center"/>
-              <StackLayout Grid.Row="3"
-                           Grid.Column="1"
-                           Orientation="Horizontal"
-                           VerticalOptions="Center">
-                     <RadioButton x:Name="FlowDirectionLTR"
-                                  AutomationId="FlowDirectionLTR"
-                                  GroupName="FlowDirection"
-                                  Content="LTR"
-                                  CheckedChanged="OnFlowDirectionChanged"/>
-                     <RadioButton x:Name="FlowDirectionRTL"
-                                  AutomationId="FlowDirectionRTL"
-                                  GroupName="FlowDirection"
-                                  Content="RTL"
-                                  CheckedChanged="OnFlowDirectionChanged"/>
-              </StackLayout>
-
-              <!-- Progress Color -->
-              <Label Grid.Row="4"
-                     Grid.Column="0"
-                     Text="Progress Color:"
-                     VerticalOptions="Center"/>
-              <StackLayout Grid.Row="4"
-                           Grid.Column="1"
-                           Orientation="Horizontal"
-                           Spacing="5"
-                           VerticalOptions="Center">
-                     <Button Text="Green"
-                             BackgroundColor="Green"
-                             TextColor="White"
-                             Clicked="ProgressColorButton_Clicked"
-                             AutomationId="ProgressColorGreenButton"/>
-                     <Button Text="Red"
-                             BackgroundColor="Red"
-                             TextColor="White"
-                             Clicked="ProgressColorButton_Clicked"
-                             AutomationId="ProgressColorRedButton"/>
-              </StackLayout>
-
-              <!-- Shadow -->
-              <Label Grid.Row="5"
-                     Grid.Column="0"
-                     Text="Shadow:"
-                     VerticalOptions="Center"/>
-              <RadioButton Grid.Row="5"
-                           Grid.Column="1"
-                           VerticalOptions="Center"
-                           AutomationId="ShadowTrueRadio"
-                           Content="True"
-                           CheckedChanged="OnShadowCheckedChanged"/>
-       </Grid>
-</ContentPage>
